dev.sh: remove files carefully
When expanding variables to pass to `rm`, make sure that the path
variable is set and fail if it is not.  This prevents an `rm` from
accidentally expanding to `rm ""/*` when the variable is unset:
https://github.com/ValveSoftware/steam-for-linux/issues/3671

Using `${FOO:?}` syntax will fail when `FOO` is unset.

+++ src/dev.sh
@@ -162,11 +162,13 @@ function package ()
 
     heading "Packaging ${agent_pkg_name}"
 
-    rm -Rf "${LAYOUT_DIR}/_diag"
+    rm -Rf "${LAYOUT_DIR:?}/_diag"
     find "${LAYOUT_DIR}/bin" -type f -name '*.pdb' -delete
+
     mkdir -p "$PACKAGE_DIR"
+    rm -Rf "${PACKAGE_DIR:?}/*"
+
     pushd "$PACKAGE_DIR" > /dev/null
-    rm -Rf *
 
     if [[ ("$CURRENT_PLATFORM" == "linux") || ("$CURRENT_PLATFORM" == "darwin") ]]; then
         tar_name="${agent_pkg_name}.tar.gz"
